Move mcts.py to MCTS class in source folder
